EOS-15751: Commit the base changes for set SW Upgrade API
